+++ src/main/java/blackjack/domain/Deck.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import java.util.LinkedList;
 import java.util.Queue;
