+++ src/test/java/blackjack/domain/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
